disable {packrat}	R Documentation

Disable the use of Packrat in a Project

Description

Disable packrat within a project, reverting to the use of standard user package libraries.

Usage

disable(project = NULL, restart = TRUE)

Arguments

project
The directory in which packrat will be disabled (defaults to the current working directory)

restart
If TRUE, restart the R session after disabling packrat.

Note

Disabling packrat for a project removes the packrat initialization code from the .Rprofile file, resulting in the use of standard user package libraries. Note that the packrat directory is not deleted, but remains unused.

To re-enable the use of packrat for a project you can call the init function.

The restart parameter will only result in a restart of R when the R environment packrat is running within makes available a restart function via getOption("restart").

[Package packrat version 0.5.0 Index]