install_local {packrat}	R Documentation

Install a Package from a Local Repository

Description

This function can be used to install a package from a local 'repository'; i.e., a directory containing package tarballs and sources.

Usage

install_local(pkgs, ..., lib = .libPaths()[1],
  repos = get_opts("local.repos"))

Arguments

pkgs
A character vector of package names.

...
Optional arguments passed to install.

lib
The library in which the package should be installed.

repos
The local repositories to search for the package names specified.

[Package packrat version 0.5.0 Index]