packify {packrat}	R Documentation

Automatically Enter Packrat Mode on Startup

Description

Install/augment the .Rprofile in a project, so that all R sessions started in this directory enter packrat mode, and use the local project library.

Usage

packify(project = NULL, quiet = FALSE)

Arguments

project
The directory in which to install the .Rprofile file.

quiet
Be chatty?

Details

It is not normally necessary to call packify directly; these files are normally installed by init. packify can be used to restore the files if they are missing (for instance, if they were not added to source control, or were accidentally removed).

You'll need to restart R in the specified directory after running packify in order to start using the private package library.

[Package packrat version 0.5.0 Index]