packrat-external {packrat}	R Documentation

Managing External Libraries

Description

These functions provide a mechanism for (temporarily) using packages outside of the packrat private library. The packages are searched within the 'default' libraries; that is, the libraries that would be available upon launching a new R session.

Usage

with_extlib(packages, expr, envir = parent.frame())

extlib(packages)

user_lib()

packrat_lib()

Arguments

packages
A set of package names (as a character vector) to load for the duration of evaluation of expr.

expr
An R expression.

envir
An environment in which the expression is evaluated.

Examples

## Not run: 
with_extlib("lattice", xyplot(1 ~ 1))

## End(Not run)

[Package packrat version 0.4.9-3 Index]