packrat-mode {packrat}	R Documentation

Packrat Mode

Description

Use these functions to switch packrat mode on and off. When within packrat mode, the R session will use the private library generated for the current project.

Usage

packrat_mode(on = NULL, project = NULL,
  auto.snapshot = get_opts("auto.snapshot"), clean.search.path = TRUE)

on(project = NULL, auto.snapshot = get_opts("auto.snapshot"),
  clean.search.path = TRUE, print.banner = TRUE)

off(project = NULL, print.banner = TRUE)

Arguments

on
Turn packrat mode on (TRUE) or off (FALSE). If omitted, packrat mode will be toggled.

project
The directory in which packrat mode is launched – this is where local libraries will be used and updated.

auto.snapshot
Perform automatic, asynchronous snapshots?

clean.search.path
Detach and unload any packages loaded from non-system libraries before entering packrat mode?

print.banner
Print the packrat banner when entering / exiting packrat mode? The packrat banner informs you of the new packrat mode state, as well as the library path in use.

[Package packrat version 0.5.0 Index]