repos_create {packrat}	R Documentation

Create a Local, CRAN-like Repository

Description

Generate a local CRAN-like repository which can be used to store and distribute R packages.

Usage

repos_create(path, name = basename(path), add = TRUE)

Arguments

path
Path to a local CRAN-like repository.

name
The name to assign to the repository. Defaults to the directory name in which the reopsitory is created.

add
Add this new repository to the current set of repositories?

[Package packrat version 0.5.0 Index]