repos_upload {packrat}	R Documentation

Upload a Package to a Local CRAN-like Repository

Description

Upload a Package to a Local CRAN-like Repository

Usage

repos_upload(package, to, ...)

Arguments

package
Path to a package tarball. The tarball should be created by R CMD build; alternatively, it can be the path to a folder containing the source code for a package (which will then be built with R CMD build) and then uploaded to the local repository.

to
The name of the CRAN-like repository. It (currently) must be a local (on-disk) CRAN repository.

...
Optional arguments passed to R CMD build.

[Package packrat version 0.5.0 Index]