repository-management {packrat}	R Documentation

Add a Repository

Description

Add a repository to the set of currently available repositories. This is effectively an easier-to-use wrapper over interacting with the "repos" option, which is otherwise set with options(repos = ...).

Usage

repos_add(..., overwrite = FALSE)

repos_add_local(..., overwrite = FALSE)

repos_set(...)

repos_set_local(...)

repos_remove(names)

repos_list()

Arguments

...
Named arguments of the form <repoName> = <pathToRepo>.

overwrite
Boolean; overwrite if a repository with the given name already exists?

names
The names of repositories (as exist in e.g. names(getOption("repos"))).

Details

repos_add_local is used for adding file-based repositories; that is, CRAN repositories that live locally on disk and not on the internet / local network.

[Package packrat version 0.5.0 Index]