status {packrat}	R Documentation

Show differences between the last snapshot and the library

Description

Shows the differences between the project's packrat dependencies, its private package library, and its R scripts.

Usage

status(project = NULL, lib.loc = libDir(project), quiet = FALSE)

Arguments

project
The directory that contains the R project.

lib.loc
The library to examine. Defaults to the private library associated with the project directory.

quiet
Print detailed information about the packrat status to the console?

Details

These differences are created when you use the normal R package management commands like install.packages, update.packages, and remove.packages. To bring these differences into packrat, you can use snapshot.

Differences can also arise if one of your collaborators adds or removes packages from the packrat dependencies. In this case, you simply need to tell packrat to update your private package library using restore.

Value

Either NULL if a packrat project has not yet been initialized, or a (invisibly) a data.frame with components:

package
The package name,

packrat.version
The package version used in the last snapshot,

packrat.source
The location from which the package was obtained,

library.version
The package version available in the local library,

currently.used
Whether the package is used in any of the R code in the current project.

[Package packrat version 0.5.0 Index]